# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MarTech Intelligence & Campaign Decision Engine
  version: 1.0.0
  description: Backend-first MarTech platform — event ingestion, customer intelligence, audience recommendation, AI campaign analysis.
servers:
  - url: /api/v1
paths:
...
  /campaigns:
    get:
      summary: List campaigns with summary metrics
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [draft, active, paused, completed]}}
        - {name: channel, in: query, schema: {$ref: '#/components/schemas/Channel'}}
        - {name: limit, in: query, schema: {type: integer}}
        - {name: cursor, in: query, schema: {type: string}}
      responses:
        '200': {description: ListResponse<Campaign>}
  /campaigns/{id}/analytics:
    get:
      summary: Campaign metrics (sends, rates, per-channel breakdown, anomalies)
      parameters: [{$ref: '#/components/parameters/externalId'}]
      responses:
        '200': {description: CampaignMetrics}
        '404': {description: Not found}
...
  /predictions/channel:
    post:
      summary: Predict the best channel for an upcoming campaign (platform, audience or customer)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChannelPredictionRequest'}
      responses:
        '200': {description: Ranked channels, content: {application/json: {schema: {$ref: '#/components/schemas/ChannelPredictionResponse'}}}}
        '400': {description: Validation failed}
        '404': {description: Unknown customer}
components:
  parameters:
    externalId:
      name: id
      in: path
      required: true
      schema: {type: string}
      description: Row external_id (e.g. cust_00042, camp_007)
  schemas:
    Channel: {type: string, enum: [email, sms, whatsapp, push, web]}
...
    ChannelPredictionRequest:
      type: object
      required: [objective]
      properties:
        objective: {type: string, enum: [conversion, engagement, retention, reactivation, awareness]}
        channels: {type: array, items: {$ref: '#/components/schemas/Channel'}, description: candidates; default all}
        customer_id: {type: string, description: "scope=customer; exclusive with audience"}
        audience:
          type: object
          description: "scope=audience; exclusive with customer_id"
          properties:
            min_score: {type: number, minimum: 0, exclusiveMaximum: 1}
            last_active_days: {type: integer, minimum: 0}
        lookback_days: {type: integer, minimum: 1, maximum: 365, default: 90}
    ChannelPredictionResponse:
      type: object
      properties:
        objective: {type: string}
        scope: {type: string, enum: [platform, audience, customer]}
        customer_id: {type: string}
        success_event: {type: string, enum: [converted, clicked, opened]}
        recommended_channel: {$ref: '#/components/schemas/Channel'}
        confidence: {type: string, enum: [high, medium, low]}
        advice: {type: string}
        channels:
          type: array
          items:
            type: object
            properties:
              rank: {type: integer}
              channel: {$ref: '#/components/schemas/Channel'}
              predicted_rate: {type: number, description: posterior mean success rate per delivered}
              interval_90: {type: array, items: {type: number}, minItems: 2, maxItems: 2}
              prob_best: {type: number, description: P(this channel has the highest true rate)}
              low_data: {type: boolean}
              opt_out_rate: {type: number}
              bounce_rate: {type: number}
              evidence:
                type: object
                properties:
                  objective_delivered: {type: integer}
                  objective_successes: {type: integer}
                  other_objectives_delivered: {type: integer}
                  other_objectives_successes: {type: integer}
                  target_delivered: {type: integer}
                  target_successes: {type: integer}
              reasons: {type: array, items: {type: string}}
        meta:
          type: object
          properties:
            lookback_days: {type: integer}
            audience_size: {type: integer}
            model: {type: string}
            samples: {type: integer}
            took_ms: {type: integer}
...